remove: SurveyAnswerService 제거(#102)
